+++ ase/calculators/calculator.py
@@ -363,7 +363,7 @@ class Calculator(object):
     'Default parameters'
 
     def __init__(self, restart=None, ignore_bad_restart_file=False, label=None,
-                 atoms=None, **kwargs):
+                 atoms=None, directory='.', **kwargs):
         """Basic calculator implementation.
 
         restart: str
@@ -372,8 +372,13 @@ class Calculator(object):
         ignore_bad_restart_file: bool
             Ignore broken or missing restart file.  By default, it is an
             error if the restart file is missing or broken.
+        directory: str
+            Working directory in which to read and write files and
+            perform calculations.
         label: str
-            Name used for all files.  May contain a directory.
+            Name used for all files.  Not supported by all calculators.
+            May contain a directory, but please use the directory parameter
+            for that instead.
         atoms: Atoms object
             Optional Atoms object to which the calculator will be
             attached.  When restarting, atoms will get its positions and
@@ -392,11 +397,15 @@ class Calculator(object):
                 else:
                     raise
 
-        self.label = None
-        self.directory = None
+        self.directory = directory
         self.prefix = None
-
-        self.set_label(label)
+        if label is not None:
+            if directory != '.' and '/' in label:
+                raise ValueError('Directory redundantly specified though '
+                                 'directory="{}" and label="{}".  '
+                                 'Please omit "/" in label.'
+                                 .format(directory, label))
+            self.set_label(label)
 
         if self.parameters is None:
             # Use default parameters if they were not read from file:
@@ -417,6 +426,40 @@ class Calculator(object):
         if not hasattr(self, 'name'):
             self.name = self.__class__.__name__.lower()
 
+    @property
+    def label(self):
+        if self.directory == '.':
+            return self.prefix
+
+        # Generally, label ~ directory/prefix
+        #
+        # We use '/' rather than os.pathsep because
+        #   1) directory/prefix does not represent any actual path
+        #   2) We want the same string to work the same on all platforms
+        if self.prefix is None:
+            return self.directory + '/'
+
+        return '{}/{}'.format(self.directory, self.prefix)
+
+    @label.setter
+    def label(self, label):
+        if label is None:
+            self.directory = '.'
+            self.prefix = None
+            return
+
+        tokens = label.rsplit('/', 1)
+        if len(tokens) == 2:
+            directory, prefix = tokens
+        else:
+            assert len(tokens) == 1
+            directory = '.'
+            prefix = tokens[0]
+        if prefix == '':
+            prefix = None
+        self.directory = directory
+        self.prefix = prefix
+
     def set_label(self, label):
         """Set label and convert label to directory and prefix.
 
@@ -424,20 +467,12 @@ class Calculator(object):
 
         * label='abc': (directory='.', prefix='abc')
         * label='dir1/abc': (directory='dir1', prefix='abc')
+        * label=None: (directory='.', prefix=None)
 
         Calculators that must write results to files with fixed names
-        can overwrite this method so that the directory is set to all
+        can override this method so that the directory is set to all
         of label."""
-
         self.label = label
-
-        if label is None:
-            self.directory = None
-            self.prefix = None
-        else:
-            self.directory, self.prefix = os.path.split(label)
-            if self.directory == '':
-                self.directory = os.curdir
 
     def get_default_parameters(self):
         return Parameters(copy.deepcopy(self.default_parameters))
@@ -736,13 +771,15 @@ class FileIOCalculator(Calculator):
                 'Please set ${} environment variable '
                 .format('ASE_' + self.name.upper() + '_COMMAND') +
                 'or supply the command keyword')
-        command = self.command.replace('PREFIX', self.prefix)
+        command = self.command
+        if 'PREFIX' in command:
+            command = command.replace('PREFIX', self.prefix)
         errorcode = subprocess.call(command, shell=True, cwd=self.directory)
 
         if errorcode:
+            path = os.path.abspath(self.directory)
             raise CalculationFailed('{} in {} returned an error: {}'
-                                    .format(self.name, self.directory,
-                                            errorcode))
+                                    .format(self.name, path, errorcode))
         self.read_results()
 
     def write_input(self, atoms, properties=None, system_changes=None):
@@ -751,7 +788,8 @@ class FileIOCalculator(Calculator):
         Call this method first in subclasses so that directories are
         created automatically."""
 
-        if self.directory != os.curdir and not os.path.isdir(self.directory):
+        absdir = os.path.abspath(self.directory)
+        if absdir != os.curdir and not os.path.isdir(self.directory):
             os.makedirs(self.directory)
 
     def read_results(self):