+++ ase/test/db/sql_db_ext_tables.py
@@ -4,7 +4,7 @@ from ase import Atoms
 from ase.test import must_raise
 import numpy as np
 
-DB_NAMES = ["test_ext_tables.db", "postgresql"]
+DB_NAMES = ["test_ext_tables.db", "postgresql", "mysql", "mariadb"]
 
 
 def get_db_name(name):
@@ -13,6 +13,16 @@ def get_db_name(name):
             name = 'postgresql://ase:ase@postgres:5432/testase'
         else:
             name = os.environ.get('ASE_TEST_POSTGRES_URL')
+    elif name == "mysql":
+        if os.environ.get('CI_PROJECT_DIR'):  # gitlab-ci
+            name = 'mysql://root:ase@mysql:3306/testase_mysql'
+        else:
+            name = os.environ.get('MYSQL_DB_URL')
+    elif name == 'mariadb':
+        if os.environ.get('CI_PROJECT_DIR'):  # gitlab-ci
+            name = 'mariadb://root:ase@mariadb:3306/testase_mysql'
+        else:
+            name = os.environ.get('MYSQL_DB_URL')
     return name
 
 
@@ -167,7 +177,7 @@ for db_name in DB_NAMES:
     if name is None:
         continue
 
-    if db_name == "postgresql":
+    if db_name in ["postgresql", "mysql", "mariadb"]:
         c = connect(name)
         c.delete([row.id for row in c.select()])
     test_create_and_delete_ext_tab(name)
@@ -176,5 +186,5 @@ for db_name in DB_NAMES:
     test_write_atoms_row(name)
     test_external_table_upon_update()
 
-    if db_name != "postgresql":
+    if db_name not in ["postgresql", "mysql", "mariadb"]:
         os.remove(name)